% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/PeakList.R
\docType{methods}
\name{addFixedWidth}
\alias{addFixedWidth}
\alias{addFixedWidth,PeakList,numeric,numeric-method}
\title{Generic method to add/update peak width}
\usage{
addFixedWidth(object, lowerWidth, upperWidth)

\S4method{addFixedWidth}{PeakList,numeric,numeric}(object, lowerWidth,
  upperWidth)
}
\arguments{
\item{object}{PeakList object}

\item{lowerWidth}{numeric}

\item{upperWidth}{numeric}
}
\value{
object PeakList with updated/new peak widths
}
\description{
This method will update current upper/lower width for all
peaks
}
\examples{
library(tofsimsData)
data(tofsimsData)
testSpectra<-reduceSpectrumResolution(object = testSpectra, everyN = 4, mode = 'keep')
testSpectra<-smootherSpline(testSpectra, stepsize = 10, spar = 0.3)
testSpectra<-smootherGolay(testSpectra, p = 3, n = 5)
testSpectra<-peakPick(testSpectra, span = 100)
testSpectra<-addFixedWidth(testSpectra, 0.2, 0.2)
plot(testSpectra, , mzRange=c(38.5,40.5), type = 'l')
}
