% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/AllGetters.R,
%   R/AllSetters.R
\docType{methods}
\name{analysis}
\alias{analysis}
\alias{analysis<-}
\alias{analysis,MassSpectra,missing-method}
\alias{analysis,MassSpectra,numeric-method}
\alias{analysis<-,MassSpectra-method}
\title{\code{analysis}, slot of \code{MassSpectra} class objects}
\usage{
analysis(object, noAccess, ...)

analysis(object) <- value

\S4method{analysis}{MassSpectra,missing}(object)

\S4method{analysis}{MassSpectra,numeric}(object, noAccess)

\S4method{analysis}{MassSpectra}(object) <- value
}
\arguments{
\item{object}{object of class MassSpectra}

\item{noAccess}{numeric access number to analysis slot}

\item{...}{additional args}

\item{value}{object to be put in analysis slot}
}
\value{
summary or content of analysis slot
}
\description{
\code{analysis}, slot of \code{MassSpectra} class objects
}
\examples{
library(tofsimsData)
data(tofsimsData)
testImage<-PCAnalysis(testImage, nComp = 3)
## obtain summary of analysis slot content
analysis(testImage)
}
\seealso{
object \code{\link{MassSpectra}} other slots \code{\link{mz}}
\code{\link{nz}} \code{\link{analysisName}} \code{\link{instrument}}
\code{\link{calibPoints}} \code{\link{calibration}}
}
