% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/AllGetters.R,
%   R/AllSetters.R
\docType{methods}
\name{analysisName}
\alias{analysisName}
\alias{analysisName<-}
\alias{analysisName,MassSpectra-method}
\alias{analysisName<-,MassSpectra-method}
\title{\code{analysisName}, slot of \code{MassSpectra} class objects}
\usage{
analysisName(object, ...)

analysisName(object) <- value

\S4method{analysisName}{MassSpectra}(object)

\S4method{analysisName}{MassSpectra}(object) <- value
}
\arguments{
\item{object}{object of class MassSpectra}

\item{...}{further args}

\item{value}{character replacement value for slot analysisName}
}
\value{
content of analysisName slot
}
\description{
\code{analysisName}, slot of \code{MassSpectra} class objects
}
\examples{
library(tofsimsData)
data(tofsimsData)
## access name of analysis
analysisName(testSpectra)
## replace name of analysis
analysisName(testSpectra) <- 'sample001_pos001_settings_default'
analysisName(testSpectra)
}
\seealso{
object \code{\link{MassSpectra}} other slots \code{\link{mz}}
\code{\link{analysis}} \code{\link{nz}} \code{\link{instrument}}
\code{\link{calibPoints}} \code{\link{calibration}}
}
