% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcomp.R
\docType{methods}
\name{baseObject,PrComp-method}
\alias{baseObject,PrComp-method}
\title{constructor for PrComp}
\usage{
\S4method{baseObject}{PrComp}(object)
}
\arguments{
\item{object}{object of class}
}
\value{
object of class PrComp
}
\description{
constructor for PrComp
}
