% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/princomp.R
\docType{methods}
\name{baseObject,PrinComp-method}
\alias{baseObject,PrinComp-method}
\title{constructor for PrinComp}
\usage{
\S4method{baseObject}{PrinComp}(object)
}
\arguments{
\item{object}{object with class}
}
\value{
object of class PrinComp
}
\description{
constructor for PrinComp
}
