% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{baseObject}
\alias{baseObject}
\title{generic accessor method baseObject}
\usage{
baseObject(object)
}
\arguments{
\item{object}{helper for prcomp and princomp wrappers}
}
\value{
baseObject
}
\description{
generic accessor method baseObject
}

