% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassImage.R
\docType{methods}
\name{binning}
\alias{binning}
\alias{binning,MassImage-method}
\title{binning}
\usage{
binning(object, binningFactor, ...)

\S4method{binning}{MassImage}(object, binningFactor = 2)
}
\arguments{
\item{object}{object of class MassImage}

\item{binningFactor}{numeric factor for binning (2, 4, etc)}

\item{...}{additional args}
}
\value{
binned object of class MassImage
}
\description{
binning
}
\details{
bining is used to reduce the resolution/size of MassImage objects. 
Optionally \code{mclapply} from the parallel package is used to speed up 
processing time.
}
\examples{
library(BiocParallel)
testImage<-MassImage('dummy')
par(mfcol=c(1,2), oma=c(0,0,0,0), mar=c(0,0,0,0))
image(testImage)
## the following param will cause to run non parallel
register(SerialParam(), default=TRUE)
testImage <- binning(testImage,binningFactor = 4)
image(testImage)
\dontrun{
library(tofsimsData)
data(tofsimsData)
par(mfcol=c(1,2), oma=c(0,0,0,0), mar=c(0,0,0,0))
image(testImage)
testImage <- binning(testImage,binningFactor = 4)
image(testImage)}
}

