% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassSpectra.R
\docType{methods}
\name{bwApply}
\alias{bwApply}
\alias{bwApply,MassSpectra,matrix-method}
\title{bwApply}
\usage{
bwApply(object, bwMatrix)

\S4method{bwApply}{MassSpectra,matrix}(object, bwMatrix)
}
\arguments{
\item{object}{object of class MassImage}

\item{bwMatrix}{matrix with boolean or numeric 1 and 0}
}
\value{
object of class MassImage multiplied with B/W matrix
}
\description{
bwApply allow to get new object from a black / white matrix
All NZs at black positions will be taken
}
\examples{
library(tofsimsData)
data(tofsimsData)
testImage <- PCAnalysis(testImage, nComp = 2)
library(EBImage)
mask<-thresh(imageMatrix(analysis(testImage,noAccess = 1),comp = 1), w = 15, h = 15)
#inverse of mask
mask <- (mask-1)^2
par(mfcol=c(1,2), oma=c(0,0,0,0), mar=c(0,0,0,0))
image(testImage)
image(bwApply(testImage, mask))
}

