% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassSpectra.R
\docType{methods}
\name{calibPointNew}
\alias{calibPointNew}
\alias{calibPointNew,MassSpectra,numeric-method}
\title{Generic method calibPointNew that modifies slot calibPoints}
\usage{
calibPointNew(object, mz, reset = FALSE, value = NULL)

\S4method{calibPointNew}{MassSpectra,numeric}(object, mz, reset = FALSE,
  value = NULL)
}
\arguments{
\item{object}{MassSpectra object}

\item{mz}{the m/z value to be specified with a TOF value}

\item{reset}{shall the list of calibration points be reset}

\item{value}{TOF value to be assigned to mz}
}
\value{
call by reference, hence MassSpectra object with new calib point

object MassSpectra with added/updated calibration points
}
\description{
Generic method calibPointNew that modifies slot calibPoints

\code{calibPointNew} is a method to set a new mass calibration point
}
\details{
\code{calibPointNew} ia a method to set a new mass calibration point. When 
\code{value} is not provided as arguemnt, the TOF for the chosen \code{mz} 
value has to be chosen interactively by mouse.
}
\examples{
library(tofsimsData)
data(tofsimsData)
testSpectra <- calibPointNew(testSpectra, mz = 15, value = 15.01551)
testSpectra <- calibPointNew(testSpectra, mz = 181, value = 181.0228)
calibPoints(testSpectra)
par(mfcol=c(1,2))
plot(testSpectra,mzRange=c(38.5,40.5),type='l')
testSpectra <- recalibrate(testSpectra)
plot(testSpectra, mzRange=c(38.5,40.5), type='l')
}
