% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/AllGetters.R, R/AllSetters.R
\docType{methods}
\name{calibPoints}
\alias{calibPoints}
\alias{calibPoints,MassSpectra-method}
\alias{calibPoints<-}
\alias{calibPoints<-,MassSpectra-method}
\title{\code{calibPoints}, slot of \code{MassSpectra} class objects}
\usage{
calibPoints(object)

calibPoints(object) <- value

\S4method{calibPoints}{MassSpectra}(object)

\S4method{calibPoints}{MassSpectra}(object) <- value
}
\arguments{
\item{object}{object of class MassSpectra}

\item{value}{data.frame replacement values for calibPoints slot}
}
\value{
contents of slot calibPoints
}
\description{
\code{calibPoints}, slot of \code{MassSpectra} class objects
}
\examples{
library(tofsimsData)
data(tofsimsData)
testSpectra<-calibPointNew(testSpectra, mz = 15, value = 15.0232)
testSpectra<-calibPointNew(testSpectra, mz = 181, value = 181.0569)
## access 'calibPoint' slot of 'MassSpectra' object
calibPoints(testSpectra)
## replacing values in the 'calibPoint' slot
calibPoints(testSpectra)[2,2]<-297000
calibPoints(testSpectra)
}
\seealso{
object \code{\link{MassSpectra}} other slots \code{\link{mz}}
\code{\link{analysis}} \code{\link{analysisName}} \code{\link{instrument}}
\code{\link{nz}} \code{\link{calibration}}
}

