% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/AllGetters.R,
%   R/AllSetters.R
\docType{methods}
\name{calibration}
\alias{calibration}
\alias{calibration<-}
\alias{calibration,MassSpectra-method}
\alias{calibration<-,MassSpectra-method}
\title{\code{calibration}, slot of \code{MassSpectra} class objects}
\usage{
calibration(object)

calibration(object) <- value

\S4method{calibration}{MassSpectra}(object)

\S4method{calibration}{MassSpectra}(object) <- value
}
\arguments{
\item{object}{object of class MassSpectra}

\item{value}{data.frame with replacement values for calibration slot}
}
\value{
content of calibration slot
}
\description{
Generic setter for slot calibration<-
}
\examples{
library(tofsimsData)
data(tofsimsData)
## access calibration slot
calibration(testSpectra)
## replacing the values of the 'calibration' slot is
## possible but it makes at the moment no sense as it
## doesn't change the actual mass calibration. The 
## 'calibration' slot is just used to store the values
## while 'recalibration' uses the values from 
## 'calibPoints' slot.
calibration(testSpectra) <- data.frame(intercept = 21420, slope = 20480)
calibration(testSpectra)
}
\seealso{
object \code{\link{MassSpectra}} other slots \code{\link{mz}}
\code{\link{analysis}} \code{\link{analysisName}} \code{\link{instrument}}
\code{\link{calibPoints}} \code{\link{nz}}
}
