% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/PeakList.R
\docType{methods}
\name{changePeakWidth}
\alias{changePeakWidth}
\alias{changePeakWidth,PeakList,missing,missing,missing-method}
\alias{changePeakWidth,PeakList,numeric,numeric,numeric-method}
\title{method changePeakWidth}
\usage{
changePeakWidth(object, selectMz, lowerWidth, upperWidth, ...)

\S4method{changePeakWidth}{PeakList,missing,missing,missing}(object, selectMz,
  lowerWidth, upperWidth, ...)

\S4method{changePeakWidth}{PeakList,numeric,numeric,numeric}(object, selectMz,
  lowerWidth, upperWidth, ...)
}
\arguments{
\item{object}{PeakList object}

\item{selectMz}{numeric change width of peak closest to selectMz}

\item{lowerWidth}{numeric lower width value in mass units}

\item{upperWidth}{numeric upper width value in mass units}

\item{...}{additional args}
}
\value{
PeakList object with upated peak widths
}
\description{
method changePeakWidth
}
\details{
method changePeakWidth is used to modify the peak width of an individual peak
it should be called with the argument mzRange to zoom into the region of
interest for choosing the peak. Then two further clicks for choosing the
(new) lower and upper peak widths.
}
\examples{
library(tofsimsData)
data(tofsimsData)
testPeakList<-PeakList(analysisName = analysisName(testSpectra),
instrument = instrument(testSpectra),
nz = nz(testSpectra),
calibration = calibration(testSpectra),
calibPoints = calibPoints(testSpectra),
mz = mz(testSpectra),
peakIDs = NULL,
peakMzs = NULL)
par(mfcol=c(1,2))
testPeakList<-addPeaks(testPeakList, mzs=26:31, width=0.4)
peakWidths(testPeakList)
testPeakList<-changePeakWidth(testPeakList, selectMz = 27, lowerWidth = 0.2, upperWidth = 0.3)
peakWidths(testPeakList)
}
