% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{check.extension}
\alias{check.extension}
\title{Check file extension}
\usage{
check.extension(filepath, extension)
}
\arguments{
\item{filepath}{character}

\item{extension}{character}
}
\value{
boolean
}
\description{
Function to check file extension
}
\details{
This function is used for check the file extension
}
\author{
Lorenz Gerber, Viet Mai Hoang
}

