% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNF.R
\name{computeMNF}
\alias{computeMNF}
\title{compute MNF}
\usage{
computeMNF(nzData = NULL, noise = NULL, SNR = NULL, ind = NULL,
  iter = TRUE, limitSNR = NULL, covNoise = NULL)
}
\arguments{
\item{nzData}{matrix}

\item{noise}{matrix}

\item{SNR}{numeric}

\item{ind}{numeric}

\item{iter}{boolean}

\item{limitSNR}{numeric}

\item{covNoise}{matrix}
}
\value{
MNF transform
}
\description{
compute MNF, helper for MNF/nnMNF
}
\details{
This is a helper function for the MNF/nnMNF function and originates
from the \code{mzImage} package.
}

