% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNF.R
\name{computeNoise}
\alias{computeNoise}
\title{computeNoise}
\usage{
computeNoise(stat, x, y)
}
\arguments{
\item{stat}{unknown}

\item{x}{unknown}

\item{y}{unknown}
}
\value{
matrix numeric noise
}
\description{
computeNoise determinates the noise by nearest neighbour estimate.
This is a helper function for the nnMNF method.
}
\details{
computeNoise determinates the noise by nearest neighbour estimate.
This is a helper function for the nnMNF method and originates 
from the \code{mzImage} package.
}

