% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassImage.R
\docType{methods}
\name{coordToPixel,MassImage,numeric-method}
\alias{coordToPixel,MassImage,numeric-method}
\title{coordToPixel}
\usage{
\S4method{coordToPixel}{MassImage,numeric}(object, xy)
}
\arguments{
\item{object}{of class MassImage}

\item{xy}{numeric vector with x/y locator coordinate}
}
\value{
xy coordinate of MassImage pixels
}
\description{
coordToPixel
}
\details{
coordToPixel translates xy coordinates from the locator() function
to cell coordinates from the image function. Origo is according to
ToF-SIMS images the upper left corner.
}

