% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAF.R
\name{covDiffCalc}
\alias{covDiffCalc}
\title{covDiffCalc calculates a x/y shift covariance matrix of a multispectral
image according to Switzer and Green 1984.}
\usage{
covDiffCalc(nzData, dataObject)
}
\arguments{
\item{nzData}{unknown}

\item{dataObject}{unknown}
}
\value{
shifted cov matrix
}
\description{
covDiffCalc calculates a x/y shift covariance matrix of a multispectral
image according to Switzer and Green 1984.
}

