% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{cReadRawPhi}
\alias{cReadRawPhi}
\title{Ulvac phi ToF-SIMS raw data import}
\usage{
cReadRawPhi(analysisName, mode = c("spectra", "imagepeaks", "image"),
  PeakListobj = c(), ...)
}
\arguments{
\item{analysisName}{character}

\item{mode}{character}

\item{PeakListobj}{object of class PeakList}

\item{...}{additional args}
}
\value{
parsed rawdata for further processing
}
\description{
Function to read raw data from the ulvac-phi trift TOF-SIMS
}
\details{
This import function works on data recorded on the ulvac-phi
trift ToF-SIMS with WinCadence software version V4.2. Other
versions mostl likley will not work. In the current version,
data has to be imported with 16bit word length, then converted
to 64bit binary and finally converted and read with the
word lenghts of the respective variables. Currently, the data
is unit mass binned with bins of size one from -0.5 to + 0.5.
}
\author{
Lorenz Gerber, Viet Mai Hoang
}
