% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ctable}
\alias{cTable}
\alias{ctable}
\title{ctable is a C++ implementation to make contingency tables}
\usage{
cTable(vect)
}
\arguments{
\item{vect}{NumericVector}
}
\value{
vars freqs
}
\description{
ctable is a C++ implementation to make contingency tables
}

