% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassImage.R
\docType{methods}
\name{dim,MassImage-method}
\alias{dim,MassImage-method}
\title{method dim for MassImage}
\usage{
\S4method{dim}{MassImage}(x)
}
\arguments{
\item{x}{object of class MassImage}
}
\value{
vector numeric
}
\description{
method dim for MassImage
}
