% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSpectra.R
\docType{methods}
\name{dim,MassSpectra-method}
\alias{dim,MassSpectra-method}
\title{method definition 'dim' for 'MassSpectra'
dim is a primitive}
\usage{
\S4method{dim}{MassSpectra}(x)
}
\arguments{
\item{x}{object object of type MassSpectra}
}
\value{
numeric value
}
\description{
method definition 'dim' for 'MassSpectra'
dim is a primitive
}
