% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EigenDecompose}
\alias{EigenDecompose}
\title{EigenDecompose for the MNF analysis}
\usage{
EigenDecompose(A, B, startIndex, endIndex)
}
\arguments{
\item{A}{NumericMatrix}

\item{B}{NumericMatric}

\item{startIndex}{int}

\item{endIndex}{int}
}
\value{
eigval eigvec mA mB
}
\description{
EigenDecompose for the MNF analysis
}
