% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{extract.header.data}
\alias{extract.header.data}
\title{extract variable names and values from Ulvac-phi
ToF-SIMS datafile headers}
\usage{
extract.header.data(header)
}
\arguments{
\item{header}{header as a raw character string}
}
\value{
list with two vectors containing variable names
and values as characters
}
\description{
Extracting the data from a Ulvac-phi Tof-SIMS
raw header character string.
}
\details{
This function takes a raw header character string read by
get.raw.header() as input and extracts variable names and values.
values are currently forwarded just as character string. This is
a helper function for read.raw.phi.
}
\author{
Lorenz Gerber
}
