% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakList.R
\name{findClosestMatch}
\alias{findClosestMatch}
\title{Find single value 'toMatch' in vector 'MatchIn'}
\usage{
findClosestMatch(toMatch, matchIn, twoMatch)
}
\arguments{
\item{toMatch}{numeric}

\item{matchIn}{vector numeric}

\item{twoMatch}{character 'upper' or 'mean'}
}
\value{
numeric ID of match
}
\description{
Find single value 'toMatch' in vector 'MatchIn'
}

