% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassSpectra.R
\docType{methods}
\name{getTOFs}
\alias{getTOFs}
\alias{getTOFs,MassSpectra-method}
\title{generic method to calculate and get TOFs}
\usage{
getTOFs(object)

\S4method{getTOFs}{MassSpectra}(object)
}
\arguments{
\item{object}{object of class MassSpectra}
}
\value{
vector with ToFs

vector numeric with TOF values
}
\description{
generic method to calculate and get TOFs

method getTOFs
}
\examples{
library(tofsimsData)
data(tofsimsData)
timeOfFlight <- getTOFs(testSpectra)
head(timeOfFlight)
}

