% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassImage.R, R/PCA.R
\name{image}
\alias{image}
\alias{image,MassImage-method}
\alias{image,PCA-method}
\title{set a generic method for image}
\usage{
image(x, ...)

\S4method{image}{MassImage}(x, ..., mzSelect = NULL)

\S4method{image}{PCA}(x, comp, ...)
}
\arguments{
\item{x}{object object with image data}

\item{...}{additional args}

\item{mzSelect}{vector, which m/z to combine for visualization. if none are 
chosen, the TIC is shownhel}

\item{comp}{numeric which component to visualize}
}
\value{
graphical output

image plot of the ToF SIMS image data
}
\description{
Method to visualize an IMS Mass Image of class \code{MassImage}
}
\examples{
testImage<-MassImage('dummy')
image(testImage)
\dontrun{
library(tofsimsData)
data(tofsimsData)
image(testImage)}
library(tofsimsData)
data(tofsimsData)
testImage<-PCAnalysis(testImage,3)
image(analysis(testImage, 1), comp = 1)
}
