% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassImage.R, R/PCA.R
\docType{methods}
\name{imageMatrix}
\alias{imageMatrix}
\alias{imageMatrix,MassImage-method}
\alias{imageMatrix,PCA-method}
\title{generic method to obtain imageMatrix}
\usage{
imageMatrix(object, ...)

\S4method{imageMatrix}{MassImage}(object)

\S4method{imageMatrix}{PCA}(object, comp, ...)
}
\arguments{
\item{object}{object of class MassImage}

\item{...}{additional args}

\item{comp}{numeric which component}
}
\value{
numeric matrix

matrix numeric
}
\description{
generic method to obtain imageMatrix

Method imageMatrix for class MassImage
}
\examples{
library(tofsimsData)
data(tofsimsData)
## the TIC matrix can be extracted 
dataMatrix <- imageMatrix(testImage)
dim(dataMatrix)
## the matrix can be visualized with the 
## normal image() function
image(dataMatrix)
}
