% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{import.raw}
\alias{import.raw}
\title{Raw data import}
\usage{
import.raw(analysisName, mode = c("spectra", "imagepeaks"),
  PeakListobj = c(), untilScan = NULL, ...)
}
\arguments{
\item{analysisName}{character}

\item{mode}{charcter}

\item{PeakListobj}{object of class PeakList}

\item{untilScan}{numeric read data up to which scan number}

\item{...}{addtional args}
}
\value{
parsed rawdata for further processing
}
\description{
Function to read raw data.
}
\details{
This import function works on GRD and ITZIP format
}
\author{
Lorenz Gerber, Viet Mai Hoang
}
