% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{import}
\alias{import}
\title{import is the C++ code for importing iontof raw data}
\usage{
import(rFilename, fType, imageSize, upperMass)
}
\arguments{
\item{rFilename}{CharacterVector}

\item{fType}{CharacterVector}

\item{imageSize}{int}

\item{upperMass}{int}
}
\value{
imported binary raw data
}
\description{
import is the C++ code for importing iontof raw data
}

