% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/AllGetters.R,
%   R/AllSetters.R
\docType{methods}
\name{instrument}
\alias{instrument}
\alias{instrument<-}
\alias{instrument,MassSpectra-method}
\alias{instrument<-,MassSpectra-method}
\title{\code{instrument}, slot of \code{MassSpectra} class objects}
\usage{
instrument(object, ...)

instrument(object) <- value

\S4method{instrument}{MassSpectra}(object)

\S4method{instrument}{MassSpectra}(object) <- value
}
\arguments{
\item{object}{object of class MassSpectra}

\item{...}{additional args}

\item{value}{character name of instrument used in the experiment}
}
\value{
content of instrument slot
}
\description{
\code{instrument}, slot of \code{MassSpectra} class objects
}
\examples{
library(tofsimsData) 
data(tofsimsData)
## access instrument slot in MassSpectra objects
instrument(testSpectra)
## values for the 'instrument' slot can currently be
## 'iontof' or 'ulvacphi'. It is not advisable to 
## change those values manually
}
\seealso{
object \code{\link{MassSpectra}} other slots \code{\link{mz}}
\code{\link{analysis}} \code{\link{analysisName}} \code{\link{nz}}
\code{\link{calibPoints}} \code{\link{calibration}}
}
