% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGetters.R
\docType{methods}
\name{iters,MCR-method}
\alias{iters,MCR-method}
\title{MCR accessor iters,}
\usage{
\S4method{iters}{MCR}(object)
}
\arguments{
\item{object}{object of class MCR}
}
\value{
iters from object
}
\description{
MCR accessor iters,
}

