% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{iters}
\alias{iters}
\title{generic accessor for iters slot}
\usage{
iters(object)
}
\arguments{
\item{object}{object of class MCR}
}
\value{
content of iters slot
}
\description{
generic accessor for iters slot
}

