% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{itzipName<-}
\alias{itzipName<-}
\title{generic for setter itzipName}
\usage{
itzipName(object) <- value
}
\arguments{
\item{object}{internal}

\item{value}{internal}
}
\value{
object with updated itzipName slot
}
\description{
generic for setter itzipName
}
