% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{itzipName}
\alias{itzipName}
\title{defining generic accessor method for "itzipName"}
\usage{
itzipName(object)
}
\arguments{
\item{object}{internal}
}
\value{
content of itzipName
}
\description{
defining generic accessor method for "itzipName"
}

