% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNF.R
\name{LapackGenEigen}
\alias{LapackGenEigen}
\title{LapackGenEigen}
\usage{
LapackGenEigen(A, B, IL = 1, IU = 3)
}
\arguments{
\item{A}{matrix}

\item{B}{matrix}

\item{IL}{int start index}

\item{IU}{int end index}
}
\value{
list with values, vectors and info
}
\description{
LapackGenEigen is helper function for MNF and nnMNF
}
\details{
LapackGenEigen is adapted from the \code{mzImage} package. While
it initially used dsygvx from the \code{LAPACK} library, it is now
ported to \code{RcppArmadillo}, using the \code{eig_pair} function.
}
