% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{legend.col}
\alias{legend.col}
\title{legend.col}
\usage{
legend.col(col, lev)
}
\arguments{
\item{col}{character color}

\item{lev}{character levels}
}
\value{
graphical output
}
\description{
legend.col is a helper for the plot function of Scoreplots. It 
allows to visualize a third component by a color range. legend.col
plots the color range as legend on the side of the plot
}
