% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{look.for.itzip.property}
\alias{look.for.itzip.property}
\title{Get ITZIP property value}
\usage{
look.for.itzip.property(itzipName, itzipProperties)
}
\arguments{
\item{itzipName}{character}

\item{itzipProperties}{character}
}
\value{
character value from itzipProperties corresponding itzipName
}
\description{
Function to extract value by passing property name
}
\details{
This function is used to get ITZIP property value by passing its name
}
\author{
Lorenz Gerber, Viet Mai Hoang
}

