% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassDefinition.R, R/MAF.R
\name{MAF}
\alias{MAF}
\alias{MAF}
\title{Class MAF}
\usage{
MAF(dataObject, nComp = 10, usePCA = TRUE)

MAF(dataObject, nComp = 10, usePCA = TRUE)
}
\arguments{
\item{dataObject}{object of type MassImage}

\item{nComp}{integer number of components}

\item{usePCA}{boolean use PCA}
}
\value{
object of type MAF
}
\description{
Class \code{MAF} contains methods for Maximum Autocorrelation Factors 
analysis

MAF is a Maximum Autocorrelation Factor Analysis
}
\details{
Class \code{MAF} contains methods for Maximum Autocorrelation Factors 
analysis

MAF is a Maximum Autocorrelation Factor Analysis. The code is implemented from the
publication of
}
\examples{
library(tofsimsData)
data(tofsimsData)
\dontrun{data(tofsimsData)
MAF(testImage,5,TRUE)
image(analysis(testImage,1),comp = 1)}
}
