% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSpectra.R
\docType{methods}
\name{makeTIC,MassSpectra-method}
\alias{makeTIC,MassSpectra-method}
\title{Method makeTIC for MassSpectra Class}
\usage{
\S4method{makeTIC}{MassSpectra}(object)
}
\arguments{
\item{object}{object of class MassSpectra}
}
\value{
object of class MassSpectra with just one spectra, the TIC
}
\description{
Method makeTIC sums up all Mass Spectra in the called Mass Spectra object
}

