% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{makeTIC}
\alias{makeTIC}
\title{generic for makeTIC}
\usage{
makeTIC(object)
}
\arguments{
\item{object}{object of type MassSpectra}
}
\value{
object of class MassSpectra with TIC
}
\description{
generic for makeTIC
}

