% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakList.R
\name{manualSelectPeaks}
\alias{manualSelectPeaks}
\title{This method is base method for plotting and manual select data}
\usage{
manualSelectPeaks(object, n = 512, ...)
}
\arguments{
\item{object}{object of type PeakList}

\item{n}{numeric}

\item{...}{additional args}
}
\value{
numeric x coordinates
}
\description{
This method is base method for plotting and manual select data
}
