% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassDefinition.R, R/MassImage.R
\name{MassImage}
\alias{MassImage}
\title{Class MassImage}
\usage{
MassImage(select = c("ulvacbif", "iontofbif", "iontofgrdpeaks",
  "ulvacrawpeaks", "dummy"), analysisName, PeakListobj = c(),
  untilScan = NULL, ...)

MassImage(select = c("ulvacbif", "iontofbif", "iontofgrdpeaks",
  "ulvacrawpeaks", "dummy"), analysisName, PeakListobj = c(),
  untilScan = NULL, ...)
}
\arguments{
\item{select}{\code{character}, 'ulvacbif', 'iontofbif',
'iontofgrdpeaks', 'ulvacrawpeaks', 'dummy'}

\item{analysisName}{\code{character}, name of analysis}

\item{PeakListobj}{\code{PeakList} class object, used as peaklist for 
rawdata import}

\item{untilScan}{integer or NULL to determine number of ToF-SIMS scans to 
import}

\item{...}{additional args}
}
\value{
object of class MassImage
}
\description{
Class MassImage contains the information to shape a number of mass spectra
into an image.

\code{MassImage}  is also the call to the class constructor. It is used for 
importing both BIF/BIF6 and raw image data.
}
\details{
Class \code{MassImage} inherits from the classes \code{MassAnalysis} and
\code{MassSpectra}. It contains the information to shape a number of mass 
spectra into an image.

\code{MassImage} is the user class constructor to obtain a MassImage object.
Data can be imported from BIF or raw data files (Iontof or Ulvacphi). To 
import raw data, a MassSpectra object with a valid PeakList object has to 
be provided as argument.
}
\section{Slots}{

\describe{
\item{\code{xy}}{vector giving the pixel dimension of the image}
}}
\examples{
# creating dummy data
testImage<-MassImage('dummy')
image(testImage)
\dontrun{
# import of rawdata
# first a PeakList object has to be created
library(tofsimsData)
data(tofsimsData)
testSpectra <- calibPointNew(testSpectra, mz = 15, value = 15.01551)
testSpectra <- calibPointNew(testSpectra, mz = 181, value = 181.0228)
testSpectra <- recalibrate(testSpectra)
testSpectra <- unitMassPeaks(testSpectra, mzRange = c(1,200), widthAt = c(15, 181), 
factor = c(0.4, 0.6), lower = c(14.97, 15.05), upper = c(180.84, 181.43))
# obtaining the path to the raw data file in 'tofsims' package
importFile<-system.file("rawdata", "trift_test_001.RAW", package = "tofsimsData")
rawImportedImage <- MassImage('ulvacrawpeaks', importFile, 
PeakListobj = testSpectra)
image(rawImportedImage)
}
}
\author{
Lorenz Gerber <lorenz.gerber@slu.se>
}

