% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassDefinition.R, R/MassSpectra.R
\name{MassSpectra}
\alias{MassSpectra}
\alias{MassSpectra}
\title{Class \code{MassSpectra}}
\usage{
MassSpectra(select = c("ulvacraw", "iontofgrd", "dummy"), analysisName, ...)

MassSpectra(select = c("ulvacraw", "iontofgrd", "dummy"), analysisName, ...)
}
\arguments{
\item{select}{\code{character}, 'ulvacraw', 'iontofgrd', 'dummy'}

\item{analysisName}{\code{character}, the (file)name of the dataset}

\item{...}{additional args}
}
\value{
object of class MassSpectra
}
\description{
Class \code{MassSpectra} is the main data container in the tofsims
package as it contains the individual mass spectra.

\code{MassSpectra} is also the call to class constructor. It is used
for importing high-resolution mass spectra from raw data.
}
\details{
Class \code{MassSpectra} is the main data container of the \code{tofsims} 
package, containing the individual mass spectra in the slot \code{nz}. 
Additional metadata about the analysis can be found in the slots 
\code{analysisName} and \code{instrument}. Values for slope and intercept 
of the linear mass calibration equation are stored in the slot 
\code{calibration}. The M/z values can be found in \code{mz}. 
\code{calibration} allows calculating from M/z values back to 
times-of-flight.
The slot \code{calibPoints} is used to recalibrate the dataset. It contains 
a data.frame with the columns \code{mz} and \code{TOF}. The slot 
\code{analysis} of type \code{list}, is used as a container for data 
analysis objects. Typically, object of the class \code{MassSpectra} are 
constructed during data import using the user constructor function with the 
same name as the class, \code{MassSpectra}.

\code{MassSpectra} is also the call to class constructor. It is used
for importing high-resolution mass spectra from raw data.
}
\section{Slots}{

\describe{
\item{\code{analysisName}}{character vector with the import filename}

\item{\code{instrument}}{character vector type of instrument used in the experiment}

\item{\code{calibration}}{data frame for numerics slope and intercept of the mass 
calibration}

\item{\code{calibPoints}}{data frame for time of flight to maass to charge 
calibration}

\item{\code{nz}}{matrix with rows of ion counts and columns as toftimes or mass to 
charge ratios}

\item{\code{mz}}{vector same length as columns in \code{nz} for mass to charge 
values}
}}

\examples{
\dontrun{
## access rawdata in tofsims package
library(tofsimsData)
importFile<-system.file("rawdata", "trift_test_001.RAW", package = "tofsimsData")
MassSpectra('ulvacraw', importFile)
}
## create dummy MassSpectra object
MassSpectra('dummy')
}
\author{
Lorenz Gerber <lorenz.gerber@slu.se>
}
