% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassDefinition.R, R/MCR.R
\docType{class}
\name{MCR-class}
\alias{MCR}
\alias{MCR-class}
\alias{opaMCR}
\title{Class MCR}
\usage{
opaMCR(dataObject, opaComps, maxiter = 10)
}
\arguments{
\item{dataObject}{object of class MassImage}

\item{opaComps}{numeric number of components for the opa method}

\item{maxiter}{numeric how many iterations}
}
\value{
object of class MCR
}
\description{
Class \code{MCR} contains methods for 'Multivariate Curve 
Resolution by Alternate Least Squares'

opaMCR is a MCR-ALS function using the Orthogonal Projection Approach from
}
\details{
Class \code{MCR} contains methods for 'Multivariate Curve
Resolution by Alternate Least Squares'

opaMCR uses the function \code{ChemometricsWithR::opa()} (Orthogonal Projection Approach,
CRAN package 'ChemometricsWithR') for start estimates of pure spectras and
\code{ALS::als()} (CRAN package 'ALS') as MCR-ALS implementation. This method is doing
fine with images up to 256x256 pixels. For larger images, memory usage 
becomes unreasonably high.
}
\section{Slots}{

\describe{
\item{\code{RSS}}{numeric residual sum of squares}

\item{\code{resids}}{matrix with residuals}

\item{\code{iters}}{numeric number of iterations}
}}
\examples{
testImage<-MassImage('dummy')
testImage<-opaMCR(testImage, 2, 2)
image(analysis(testImage,1), comp = 1)
\dontrun{
library(tofsimsData)
data(tofsimsData)
testImage<-MCR(testImage, 5, 5)
image(analysis(testImage,1), comp = 1)
}
}
\author{
Lorenz Gerber <lorenz.gerber@slu.se>
}

