% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassDefinition.R, R/MNF.R
\name{MNF}
\alias{MNF}
\alias{MNF}
\title{Class MNF}
\usage{
MNF(dataObject)

MNF(dataObject)
}
\arguments{
\item{dataObject}{object of type massImage}
}
\value{
object of class MNF
}
\description{
Class \code{MNF} contains methods for Maximum Autocorrelation Factors 
analysis

This method calculates MNF transform using the 
diagonal shift method from Switzer and Green (1984)
to estimate the noise.
}
\details{
Class \code{MNF} contains methods for Maximum Autocorrelation Factors 
analysis

Minimum Noise Fraction according Green et al. (1988) using
diagonal shift method from Switzer and Green (1984) to estimate 
the noise. As the original package \code{mzImage} from 
Stone et al. 2012 is no longer maintained, we use it as code 
base for the present version. The C code was implemented 
through Rcpp (Eddelbuettel and Francois, 2011). Practically,
this method uses \code{covDiffCalc} from the MAF method. The
present function is a user constructur that will create a new
analysis slot in the chosen MassSpectra/MassImage object.
}
\examples{
testImage<-MassImage('dummy')
testImage<-MNF(testImage)
image(analysis(testImage,1), comp = 1)
\dontrun{
library(tofsimsData)
data(tofsimsData)
MNF(testImage)
image(analysis(testImage,1), comp = 1)}
}
