% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGetters.R, R/AllSetters.R
\docType{methods}
\name{mz,MassSpectra-method}
\alias{mz,MassSpectra-method}
\alias{mz<-,MassSpectra-method}
\title{mz getter method}
\usage{
\S4method{mz}{MassSpectra}(object)

\S4method{mz}{MassSpectra}(object) <- value
}
\arguments{
\item{object}{of type MassSpectra}

\item{value}{double mass to charge ratio}
}
\value{
MassSpectra object with updated mz slot
}
\description{
mz getter method

mz setter method
}
\examples{
library(tofsimsData)
data(tofsimsData)
## access the mz values fo each spectra point
mz(testSpectra)[1:100]
## replace a mz value
mz(testSpectra)[1] <- 0.000025
mz(testSpectra)[1:100]
}
