% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/AllGetters.R
\docType{methods}
\name{nComp}
\alias{nComp}
\alias{nComp,PCA-method}
\title{generic accessor method for slot nComp}
\usage{
nComp(object)

\S4method{nComp}{PCA}(object)
}
\arguments{
\item{object}{object of class PCA}
}
\value{
contents of nComp slot

nuemric number of components
}
\description{
generic accessor method for slot nComp

PCA accessor \code{nComp}, number of component
}
\examples{
library(tofsimsData) 
data(tofsimsData)
testImage<-PCAnalysis(testImage,4)
nComp(analysis(testImage,1))
}

