% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSpectra.R
\docType{methods}
\name{ndim,MassSpectra-method}
\alias{ndim,MassSpectra-method}
\title{method definition 'ndim' on 'MassSpectra'}
\usage{
\S4method{ndim}{MassSpectra}(object)
}
\arguments{
\item{object}{object of type MassSpectra}
}
\value{
numeric value
}
\description{
method definition 'ndim' on 'MassSpectra'
}

