% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{ndim}
\alias{ndim}
\title{generic accessor method for slot ndim}
\usage{
ndim(object)
}
\arguments{
\item{object}{object of class MassSpectra}
}
\value{
contents of slot ndim
}
\description{
generic accessor method for slot ndim
}
