% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNF.R
\name{nearestNeighbourMean}
\alias{nearestNeighbourMean}
\title{nearestNeighbourMean}
\usage{
nearestNeighbourMean(x)
}
\arguments{
\item{x}{unknown see mzimage}
}
\value{
matrix numeric nearest neighbours
}
\description{
nearestNeighbourMean helper for nnMNF
}
\details{
function from mzimage to calculate nearest neighbour means
}
