% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nnMean}
\alias{nnMean}
\title{nnMean is C++ code for calculating nearest neighbour means in a 
2D matrix}
\usage{
nnMean(y, nrows, ncols)
}
\arguments{
\item{y}{NumericVector}

\item{nrows}{int}

\item{ncols}{int}
}
\value{
eY
}
\description{
nnMean is C++ code for calculating nearest neighbour means in a 
2D matrix
}
