% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassDefinition.R, R/MNF.R
\name{nnMNF}
\alias{nnMNF}
\alias{nnMNF}
\title{Class nnMNF}
\usage{
nnMNF(dataObject, limitSNR = 1.5)

nnMNF(dataObject, limitSNR = 1.5)
}
\arguments{
\item{dataObject}{object of type MassImage}

\item{limitSNR}{numeric}
}
\value{
object of class MNF
}
\description{
Class \code{nnMNF} contains methods for Maximum Autocorrelation Factors 
analysis

This method calculates MNF transform using an 
\code{nearest neighbour} estimate as implemented 
in \code{mzImage} from Stone et al. (2012).
}
\details{
Class \code{nnMNF} contains methods for Maximum Autocorrelation Factors 
analysis

Minimum Noise Fraction according Green et al. (1988) but
using a nearest neighbour estimate for the noise determination
as seen in the package \code{mzImage} from Stone et al. (2012).
As the mentioned package is no longer maintained, we used an 
archieved version as code base for a new version. The C code was
implemented through Rcpp (Eddelbuettel and Francois, 2011). The
present function is a user constructur that will create a new
analysis slot in the chosen MassSpectra/MassImage object.
}
\examples{
testImage<-MassImage('dummy')
testImage<-MNF(testImage)
image(analysis(testImage,1), comp = 1)
\dontrun{
library(tofsimsData)
data(tofsimsData)
testImage<-nnMNF(testImage)
image(analysis(testImage,1), comp = 1)}
}
