% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\docType{methods}
\name{noPlottingData,PCA-method}
\alias{noPlottingData,PCA-method}
\title{Check NULL PCA object}
\usage{
\S4method{noPlottingData}{PCA}(object)
}
\arguments{
\item{object}{object of class PCA}
}
\value{
boolean validity check of class PCA object
}
\description{
Check NULL PCA object
}
