% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{noPlottingData}
\alias{noPlottingData}
\title{generic method for 'noPlottingData' aka 'is.null'}
\usage{
noPlottingData(object)
}
\arguments{
\item{object}{object of class PCA}
}
\value{
boolean validity check of PCA object
}
\description{
generic method for 'noPlottingData' aka 'is.null'
}
