% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/AllGetters.R
\docType{methods}
\name{nPeaks}
\alias{nPeaks}
\alias{nPeaks,PeakList-method}
\title{generic method for nPeaks}
\usage{
nPeaks(object)

\S4method{nPeaks}{PeakList}(object)
}
\arguments{
\item{object}{object of class PeakList}
}
\value{
integer value for number of peaks
}
\description{
generic method for nPeaks

nPeaks accessor/getter \code{nPeaks} for PeakList Class
}
\examples{
library(tofsimsData) 
data(tofsimsData)
testSpectra<-calibPointNew(testSpectra, mz = 15, value = 15.01551)
testSpectra<-calibPointNew(testSpectra, mz = 181, value = 181.0228)
testSpectra<-recalibrate(testSpectra)
testSpectra<-unitMassPeaks(testSpectra, mzRange = c(1,200), widthAt = c(15, 181), 
factor = c(0.4, 0.6), lower = c(14.97, 15.05), upper = c(180.84, 181.43))
nPeaks(testSpectra)
}

