% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassSpectra.R
\docType{methods}
\name{overlayPlot}
\alias{overlayPlot}
\alias{overlayPlot,list-method}
\title{generic overlayPlot}
\usage{
overlayPlot(objectList, ...)

\S4method{overlayPlot}{list}(objectList, ..., type = "l", mzRange = c(1,
  200), PeakListObj = NULL, cex.legend = 0.5)
}
\arguments{
\item{objectList}{list with object of type MassSpectra}

\item{...}{additional args}

\item{type}{character type of plot, usually 'l'}

\item{mzRange}{vector numeric lower and upper range for plotting the spectra}

\item{PeakListObj}{object a PeakList object can be provided to plot peaks}

\item{cex.legend}{numeric text size}
}
\value{
graphical output

graphical output
}
\description{
generic overlayPlot

This function takes as input a list with objects of type MassSpectra. The
easiest way to obtain the input data, is to use mclapply from the parallel 
package.
}
\examples{
library(tofsimsData)
data('tofsimsData')
overlayPlot(list(testImage, testSpectra))
}
\author{
Lorenz Gerber \email{lorenz.gerber@slu.se}
}

