% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{parIndicesSearch}
\alias{parIndicesSearch}
\title{helper function for parallel processing in rawdata import routines}
\usage{
parIndicesSearch(rawVector, mzs, mzsOrder, startOrEnd = "start")
}
\arguments{
\item{rawVector}{unknown}

\item{mzs}{unknown}

\item{mzsOrder}{unknown}

\item{startOrEnd}{character 'start' or 'end'}
}
\value{
numeric indicies of time of flight
}
\description{
helper function for parallel processing in rawdata import routines
}
