% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassDefinition.R
\docType{class}
\name{PCA-class}
\alias{PCA}
\alias{PCA-class}
\title{Class PCA}
\description{
Class \code{PCA} is a virtual class for PCA that will be inherited.
}
\details{
Class \code{PCA} is a virtual class for PCA that will be inherited.
}
\section{Slots}{

\describe{
\item{\code{pcaLoadings}}{matrix that holds the loadings of a principal component 
like analysis}

\item{\code{pcaScores}}{matrix that holds the scores of a principal component like 
analysis}

\item{\code{nComp}}{numeric number of components in the principal component like 
analysis}

\item{\code{imageDim}}{vector x and y values of the image dimension}

\item{\code{classOfData}}{character a more detailed description of the analysis type}
}}

