% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/AllGetters.R
\docType{methods}
\name{pcaLoadings}
\alias{pcaLoadings}
\alias{pcaLoadings,PCA,missing-method}
\alias{pcaLoadings,PCA,numeric-method}
\title{generic accessor for slot pcaLoadings}
\usage{
pcaLoadings(object, comps = c(1, 2))

\S4method{pcaLoadings}{PCA,missing}(object)

\S4method{pcaLoadings}{PCA,numeric}(object, comps = c(1, 2))
}
\arguments{
\item{object}{object of class PCA}

\item{comps}{numeric number of components}
}
\value{
contents of slot pcaLoadings

matrix numeric with loadings

vector or matrix numeric with loadings according comps
}
\description{
generic accessor for slot pcaLoadings

PCA accessor \code{pcaLoadings}, loading matrix

PCA accessor \code{pcaLoadings}, loading matrix
}
\examples{
library(tofsimsData)
data(tofsimsData)
testImage<-PCAnalysis(testImage,4)
plot(pcaLoadings(analysis(testImage,1), comps = c(1,2)))
}

