% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAF.R
\name{pcaMAF}
\alias{pcaMAF}
\title{helper function for MAF calculation}
\usage{
pcaMAF(X, nComp)
}
\arguments{
\item{X}{matrix numeric, matrix to calculate PCA from}

\item{nComp}{number of components}
}
\value{
principal component analysis
}
\description{
helper function for MAF calculation
}

