% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassDefinition.R, R/PCAnalysis.R
\name{PCAnalysis}
\alias{PCAnalysis}
\alias{PCAnalysis}
\title{Class PCAnalysis}
\usage{
PCAnalysis(dataObject, nComp, ...)

PCAnalysis(dataObject, nComp, ...)
}
\arguments{
\item{dataObject}{object of type MassImage}

\item{nComp}{integer number of components}

\item{...}{further args}
}
\value{
PCAnalysis class object
}
\description{
Class \code{PCAnalysis} contains methods for simple PCA analyis

PCAnalysis is a PCA constructor function
}
\details{
Class \code{PCAnalysis} contains methods for simple PCA analysis

PCAnalysis constructor function uses call by reference. The new object is 
put into the \code{analysis} slot of the dataObject on which PCA was 
calculated.
}
\examples{
testImage<-MassImage('dummy')
testImage<-PCAnalysis(testImage, 4)
image(analysis(testImage, 1), comp = 1)
\dontrun{
library(tofsimsData)
data(tofsimsData)
testImage<-PCAnalysis(testImage, nComp = 4)
image(analysis(testImage, 1), comp = 1)}
}
\author{
Lorenz Gerber <lorenz.gerber@slu.se>
}
