% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/AllGetters.R
\docType{methods}
\name{pcaScores}
\alias{pcaScores}
\alias{pcaScores,PCA,ANY-method}
\alias{pcaScores,PCA,numeric-method}
\title{generic accessor for slot pcaScores}
\usage{
pcaScores(object, comps = c(1, 2))

\S4method{pcaScores}{PCA,ANY}(object)

\S4method{pcaScores}{PCA,numeric}(object, comps = c(1, 2))
}
\arguments{
\item{object}{object of class PCA}

\item{comps}{numeric number of components}
}
\value{
contents of slot pcaScores

vector or matrix numeric with scores according comps
}
\description{
generic accessor for slot pcaScores

PCA accessor \code{pcaScores}, pcaScores matrix

PCA accessor \code{pcaScores}, pcaScores matrix
}
\examples{
library(tofsimsData)
data(tofsimsData)
testImage<-PCAnalysis(testImage,4)
plot(pcaScores(analysis(testImage,1), comps = c(1,2)))
}

