% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/AllGetters.R,
%   R/AllSetters.R
\docType{methods}
\name{peakIDs}
\alias{peakIDs}
\alias{peakIDs<-}
\alias{peakIDs,PeakList-method}
\alias{peakIDs<-,PeakList-method}
\title{\code{peakIDs}, slot of \code{PeakList} class objects}
\usage{
peakIDs(object)

peakIDs(object) <- value

\S4method{peakIDs}{PeakList}(object)

\S4method{peakIDs}{PeakList}(object) <- value
}
\arguments{
\item{object}{object of class PeakList}

\item{value}{data.frame}
}
\value{
content of slot peakIDs
}
\description{
\code{peakIDs}, slot of \code{PeakList} class objects
}
\examples{
library(tofsimsData)
data(tofsimsData)
testSpectra<-calibPointNew(testSpectra, mz = 15, value = 15.01551)
testSpectra<-calibPointNew(testSpectra, mz = 181, value = 181.0228)
testSpectra<-recalibrate(testSpectra)
testSpectra<-unitMassPeaks(testSpectra, mzRange = c(1,200), widthAt = c(15, 181), 
factor = c(0.4, 0.6), lower = c(14.97, 15.05), upper = c(180.84, 181.43))
peakIDs(testSpectra)[,1:10]
}
