% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/AllGetters.R, R/AllSetters.R
\docType{methods}
\name{peakMzs}
\alias{peakMzs}
\alias{peakMzs,PeakList-method}
\alias{peakMzs<-}
\alias{peakMzs<-,PeakList-method}
\title{\code{peakMzs}, slot of \code{PeakList} class objects}
\usage{
peakMzs(object)

peakMzs(object) <- value

\S4method{peakMzs}{PeakList}(object)

\S4method{peakMzs}{PeakList}(object) <- value
}
\arguments{
\item{object}{object of class PeakList}

\item{value}{data.frame}
}
\value{
contents of slot peakMzs
}
\description{
\code{peakMzs}, slot of \code{PeakList} class objects
}
\examples{
library(tofsimsData)
data(tofsimsData)
testSpectra<-calibPointNew(testSpectra, mz = 15, value = 15.01551)
testSpectra<-calibPointNew(testSpectra, mz = 181, value = 181.0228)
testSpectra<-recalibrate(testSpectra)
testSpectra<-unitMassPeaks(testSpectra, mzRange = c(1,200), widthAt = c(15, 181), 
factor = c(0.4, 0.6), lower = c(14.97, 15.05), upper = c(180.84, 181.43))
peakMzs(testSpectra)[,1:10]
}

