% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/PeakList.R
\docType{methods}
\name{peaks2Spectra}
\alias{peaks2Spectra}
\alias{peaks2Spectra,PeakList,MassSpectra-method}
\title{generic method peaks2Spectra}
\usage{
peaks2Spectra(objectPeaks, objectSpectra)

\S4method{peaks2Spectra}{PeakList,MassSpectra}(objectPeaks, objectSpectra)
}
\arguments{
\item{objectPeaks}{object object of class PeakList}

\item{objectSpectra}{object object of class MassSpectra}
}
\value{
object of class PeakList
}
\description{
peaks2Spectra allows to transfer the peaks from a PeakList object onto
a MassSpectra object. By this, the MassSpectra object is promoted into
a PeakList object
}
\examples{
library(tofsimsData)
data(tofsimsData)
testSpectra<-reduceSpectrumResolution(testSpectra,everyN = 4, mode = 'keep')
peakPickSpectra<-testSpectra
peakPickSpectra<-calibPointNew(peakPickSpectra, mz = 15, value = 15.01551)
peakPickSpectra<-calibPointNew(peakPickSpectra, mz = 181, value = 181.0228)
peakPickSpectra<-recalibrate(peakPickSpectra)
peakPickSpectra<-unitMassPeaks(peakPickSpectra, mzRange = c(1,200), widthAt = c(15, 181), 
factor = c(0.4, 0.6), lower = c(14.97, 15.05), upper = c(180.84, 181.43))
par(mfcol = c(1,2))
plot(testSpectra, mzRange = c(38.5, 40.5), type = 'l')
testSpectra<-peaks2Spectra(peakPickSpectra, testSpectra)
plot(testSpectra, mzRange = c(38.5, 40.5), type = 'l')
}

