% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/PeakList.R
\docType{methods}
\name{peakWidths}
\alias{peakWidths}
\alias{peakWidths,PeakList-method}
\title{Generic method peakWidths}
\usage{
peakWidths(object, plot = FALSE)

\S4method{peakWidths}{PeakList}(object, plot = FALSE)
}
\arguments{
\item{object}{PeakList object}

\item{plot}{boolean should there be graphical output}
}
\value{
vector of peak widths
}
\description{
Generic method peakWidths

peakWidths
}
\details{
This method will calculate peak widths (m/z) based on
lower and upper widths.

Method to return the peakWidth values of all peaks. On plot=TRUE the
width values are ploted against the M/z of the corresponding peak.
}
\examples{
library(tofsimsData)
data(tofsimsData)
testPeakList<-PeakList(analysisName = analysisName(testSpectra),
instrument = instrument(testSpectra),
nz = nz(testSpectra),
calibration = calibration(testSpectra),
calibPoints = calibPoints(testSpectra),
mz = mz(testSpectra),
peakIDs = NULL,
peakMzs = NULL)
testPeakList<-addPeaks(testPeakList, mzs=26:31, width=0.4)
testPeakList<-findPeakWidth(testPeakList, p = 3, n = 199, 
span = 100, widthExtLower = 2, widthExtUpper = 2)
testPeakList<-peakWidths(testPeakList, plot = FALSE)
}

