% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassImage.R
\docType{methods}
\name{plot,MassImage,missing-method}
\alias{plot,MassImage,missing-method}
\title{Method \code{plot()} for \code{MassImage}}
\usage{
\S4method{plot}{MassImage,missing}(x, y, ..., mzRange = c(0, 200),
  normalize = FALSE)
}
\arguments{
\item{x}{object of type MassImage}

\item{y}{missing}

\item{...}{additional args}

\item{mzRange}{vector or lenght two, indicating the mz range to be plotted}

\item{normalize}{should the mass spectra be normalized}
}
\value{
scatter plot with loading or scores
}
\description{
Method defining \code{plot()} for the \code{MassImage} class
plot has no generic by default
}
\details{
This method will call \code{plot} method of \code{MassSpectra} class.
}
