% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassSpectra.R, R/PCA.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,MassSpectra,missing-method}
\alias{plot,PCA,ANY-method}
\title{Generic method for plot}
\usage{
plot(x, y, ...)

\S4method{plot}{MassSpectra,missing}(x, y, ..., mzRange = c(0, 200),
  normalize = FALSE)

\S4method{plot}{PCA,ANY}(x, ..., comps = c(1, 2), pcType = "pcaLoadings",
  label = FALSE, labelThreshold = 1)
}
\arguments{
\item{x}{object of type MassSpectra}

\item{y}{missing}

\item{...}{further args}

\item{mzRange}{vector or lenght two, indicating the mz range to be plotted}

\item{normalize}{boolean should the mass spectra be normalized}

\item{comps}{numeric vector of length two denominating the components to be 
plotted}

\item{pcType}{character 'pcaLoadings' or pcaScores'}

\item{label}{boolean plot label}

\item{labelThreshold}{numeric threshhold on which values to plot a label}
}
\value{
graphical output

plot of mass spectra

scatter loading/score plot
}
\description{
Generic method for plot

Method defining \code{plot()} for the \code{MassSpectra} class
plot has no generic by default
}
\details{
The output of this method is adapted for plotting mass spectra. Uncalibrated
data is plotted as xy plot while uncalibrated data is plotted as barplot. 
The parameter \code{mzRange} allows choosing the plot range directly 
according to the mz number (when calibrated). The argument \code{lineplot}, 
TRUE by default, allows to switch between line and barplot.
}
\examples{
## plot method for MassSpectra objects
library(tofsimsData)
data(tofsimsData)
plot(testSpectra, mzRange=c(1,300),type='l')
}

