% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassSpectra.R
\docType{methods}
\name{poissonScaling}
\alias{poissonScaling}
\alias{poissonScaling,MassSpectra-method}
\title{generic method for "poissonScaling"}
\usage{
poissonScaling(object, offset = 1, ...)

\S4method{poissonScaling}{MassSpectra}(object, offset = 1, ...)
}
\arguments{
\item{object}{object of class MassSpectra}

\item{offset}{numeric value for poisson scaling}

\item{...}{further args}
}
\value{
object of class MassSpectra with poission 
scaled mass spectra in slot nz

object of class MassSpectra
}
\description{
generic method for "poissonScaling"

Possion scaling for data matrices.
}
\details{
Possion scaling is proposed as the method of choice for ToF-SIMS data
see Keenan and Kotula 2004. This implementation was done according to
a description in \code{Multivariate Analysis of SIMS spectra in
ToF-SIMS: Materials Analysis by Mass Spectrometry, Vickerman and
Briggs 2013} and the \code{eigenvector wiki}. The offset is described in
the \code{eigenvector wiki}.
}
\examples{
## poisson scaling of MassSpectra objects
testImage <- MassImage('dummy')
testImage <- poissonScaling(testImage)
\dontrun{
# poission scaling on real data
library(tofsimsData)
data(tofsimsData)
par(mfcol=c(2,2))
plot(testImage,type='l')
image(testImage)
testImage <- poissonScaling(testImage)
plot(testImage,type='l')
image(testImage)
}
}
\author{
Lorenz Gerber \email{lorenz.gerber@slu.se}
}

