% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassDefinition.R, R/prcomp.R
\docType{class}
\name{PrComp-class}
\alias{PrComp-class}
\alias{PrComp}
\alias{prComp}
\title{Class PrComp}
\usage{
prComp(dataObject, ...)
}
\arguments{
\item{dataObject}{object of class MassSpectra}

\item{...}{additional args for prcomp}
}
\value{
object of class PrComp
}
\description{
Class \code{PrComp} is a wrapper for the S3 function prcomp

PrComp is a PCA constructor function
}
\details{
Class \code{PrComp} is a wrapper for the S3 function prcomp

PrComp constructor function uses call by reference. The new object is put
into the \code{analysis} slot of the dataObject on which PCA was calculated.
}
\section{Slots}{

\describe{
\item{\code{scale}}{logical see description of \code{stats::prcomp}}

\item{\code{center}}{vector see description of \code{stats::prcomp}}

\item{\code{sdev}}{vector see description of \code{stats::prcomp}}
}}

\examples{
testImage<-MassImage('dummy')
testImage<-prComp(testImage)
image(analysis(testImage, 1), comp = 1)
\dontrun{
library(tofsimsData)
data(tofsimsData)
testImage<-prComp(testImage)
image(analysis(testImage, 1), comp = 1)}
}
\author{
Lorenz Gerber <lorenz.gerber@slu.se>
}
