% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassSpectra.R
\docType{methods}
\name{recalibrate}
\alias{recalibrate}
\alias{recalibrate,MassSpectra-method}
\title{Generic method recalibrate}
\usage{
recalibrate(object)

\S4method{recalibrate}{MassSpectra}(object)
}
\arguments{
\item{object}{object of class MassSpctra}
}
\value{
object of class MassSpectra, recalibrated using the data from
slots calibPoints

object of class MassSpectra, recalibrated mass values
}
\description{
Generic method recalibrate

method recalibrate
}
\examples{
library(tofsimsData)
data(tofsimsData)
testSpectra <- calibPointNew(testSpectra, mz = 15, value = 15.01551)
testSpectra <- calibPointNew(testSpectra, mz = 181, value = 181.0228)
calibPoints(testSpectra)
par(mfcol=c(1,2))
plot(testSpectra,mzRange=c(38.5,40.5),type='l')
testSpectra <- recalibrate(testSpectra)
plot(testSpectra, mzRange=c(38.5,40.5), type='l')
}
