% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassSpectra.R
\docType{methods}
\name{reduceSpectrumResolution}
\alias{reduceSpectrumResolution}
\alias{reduceSpectrumResolution,MassSpectra-method}
\title{generic method reduceSpectrumResolution}
\usage{
reduceSpectrumResolution(object, everyN = 2, mode = "remove")

\S4method{reduceSpectrumResolution}{MassSpectra}(object, everyN = 2,
  mode = "remove")
}
\arguments{
\item{object}{object of class MassSpectra}

\item{everyN}{numeric act on every nth spectra point}

\item{mode}{character 'remove' or 'keep'}
}
\value{
object of class MassSpectra with 
reduced spectral resolution

object of class MassSpectra
}
\description{
generic method reduceSpectrumResolution

reduceSpectrumResolution
}
\details{
The method reduceSpectrumResolution for MassSpectra is used
sometimes for performance reasons.
}
\examples{
library(tofsimsData) 
data(tofsimsData)
par(mfcol=c(1,2))
plot(testSpectra,mzRange = c(40,50),type='l')
testSpectra <- reduceSpectrumResolution(object = testSpectra, everyN = 2, mode = 'remove')
plot(testSpectra, mzRange = c(40,50), type='l')
}
