% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/PeakList.R
\docType{methods}
\name{removePeaks}
\alias{removePeaks}
\alias{removePeaks,PeakList,missing,character,numeric,missing-method}
\alias{removePeaks,PeakList,missing,missing,missing,numeric-method}
\alias{removePeaks,PeakList,missing,missing,numeric,missing-method}
\alias{removePeaks,PeakList,numeric,missing,missing,missing-method}
\title{generic method removePeaks}
\usage{
removePeaks(object, mzs, operator, limit, nLocator, ...)

\S4method{removePeaks}{PeakList,missing,missing,numeric,missing}(object, mzs,
  operator, limit, nLocator, ...)

\S4method{removePeaks}{PeakList,missing,missing,missing,numeric}(object, mzs,
  operator, limit, nLocator, ...)

\S4method{removePeaks}{PeakList,numeric,missing,missing,missing}(object, mzs,
  operator, limit, nLocator, ...)

\S4method{removePeaks}{PeakList,missing,character,numeric,missing}(object, mzs,
  operator, limit, nLocator, ...)
}
\arguments{
\item{object}{object of class PeakList}

\item{mzs}{M/z's of peaks to be removed}

\item{operator}{Accept ">", "<", "==", "<=", ">=", "!="}

\item{limit}{numeric limit for peaks to be removed}

\item{nLocator}{numeric how many peaks to remove with visual selection}

\item{...}{additional args}
}
\value{
object of class PeakList with removed/updated peaks
}
\description{
generic method removePeaks

removePeaks for PeakList Class allows removing peaks below a certain treshhold of ioncounts.
the threshhold is not calculated as area, but just from the peak height (ion count at peak
center)

removePeaks for PeakList Class allows removing peaks manually

removePeaks for PeakList Class allows removing peaks manually
}
\examples{
library(tofsimsData)
data(tofsimsData)
testPeakList<-PeakList(analysisName = analysisName(testSpectra),
instrument = instrument(testSpectra),
nz = nz(testSpectra),
calibration = calibration(testSpectra),
calibPoints = calibPoints(testSpectra),
mz = mz(testSpectra),
peakIDs = NULL,
peakMzs = NULL)
par(mfcol=c(1,2))
testPeakList<-addPeaks(testPeakList, mzs = 26:31, width=0.4)
plot(testPeakList, mzRange = c(25,32), type = 'l')
testPeakList<-removePeaks(testPeakList, mzs = 27)
plot(testPeakList, mzRange = c(25,32), type = 'l')
}

