% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{resids}
\alias{resids}
\title{generic accessor method for resids}
\usage{
resids(object)
}
\arguments{
\item{object}{object of class MCR}
}
\value{
content of slot resids
}
\description{
generic accessor method for resids
}
