% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGetters.R
\docType{methods}
\name{RSS,MCR-method}
\alias{RSS,MCR-method}
\title{MCR accessor RSS,}
\usage{
\S4method{RSS}{MCR}(object)
}
\arguments{
\item{object}{object of type MCR}
}
\value{
RSS from object
}
\description{
MCR accessor RSS,
}

