% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{RSS}
\alias{RSS}
\title{generic accessor for RSS}
\usage{
RSS(object)
}
\arguments{
\item{object}{object of class MCR}
}
\value{
content of slot RSS
}
\description{
generic accessor for RSS
}

