% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassImage.R
\docType{methods}
\name{show,MassImage-method}
\alias{show,MassImage-method}
\title{method definition 'show' on 'MassImage'
show has a generic by default}
\usage{
\S4method{show}{MassImage}(object)
}
\arguments{
\item{object}{object of class MassImage}
}
\value{
data.frame character
}
\description{
method definition 'show' on 'MassImage'
show has a generic by default
}
