% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSpectra.R
\docType{methods}
\name{show,MassSpectra-method}
\alias{show,MassSpectra-method}
\title{method defining \code{show()} for the \code{MassSpectra} class
show has a generic by default}
\usage{
\S4method{show}{MassSpectra}(object)
}
\arguments{
\item{object}{object of class MassSpectra}
}
\value{
data.frame character
}
\description{
method defining \code{show()} for the \code{MassSpectra} class
show has a generic by default
}
