% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakList.R
\docType{methods}
\name{show,PeakList-method}
\alias{show,PeakList-method}
\title{method defining \code{show()} for the \code{MassSpectra} class
show has a generic by default}
\usage{
\S4method{show}{PeakList}(object)
}
\arguments{
\item{object}{object of class PeakList}
}
\value{
data.frame character
}
\description{
method defining \code{show()} for the \code{MassSpectra} class
show has a generic by default
}
