% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassSpectra.R
\docType{methods}
\name{smootherGolay}
\alias{smootherGolay}
\alias{smootherGolay,MassSpectra-method}
\title{generic method smootherGolay}
\usage{
smootherGolay(object, p = 3, n = 5, ...)

\S4method{smootherGolay}{MassSpectra}(object, p = 3, n = 5, ...)
}
\arguments{
\item{object}{object of class MassSpectra}

\item{p}{numeric parameter for savitzky-golay filter}

\item{n}{numeric parameter for savitzky-golay filter}

\item{...}{additional args}
}
\value{
object of class MassSpectra with updated mass spectra

object of class MassSpectra with smoothed TIC
}
\description{
generic method smootherGolay

Method smootherGolay for MassSpectra class
}
\examples{
library(tofsimsData)
data(tofsimsData)
testSpectraSmooth <- smootherGolay(testSpectra, p = 3, n = 9)
overlayPlot(list(testSpectra, testSpectraSmooth), mzRange = c(38.5, 40.5), type = 'l')
}

