% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassSpectra.R
\docType{methods}
\name{smootherSpline}
\alias{smootherSpline}
\alias{smootherSpline,MassSpectra-method}
\title{generic smootherSpline}
\usage{
smootherSpline(object, stepsize = 5, spar = 0.3, ...)

\S4method{smootherSpline}{MassSpectra}(object, stepsize = 5, spar = 0.3,
  ...)
}
\arguments{
\item{object}{MassSpectra}

\item{stepsize}{numeric arg for spline smoother}

\item{spar}{numeric arg for spline smoother}

\item{...}{additional args}
}
\value{
object of class MassSpectra with updated mass spectra

object of class MassSpectra
}
\description{
generic smootherSpline

method smootherSpline for TIC
}
\examples{
library(tofsimsData) 
data(tofsimsData)
testSpectraSmooth <- smootherSpline(testSpectra)
overlayPlot(list(testSpectra, testSpectraSmooth), mzRange = c(38.5, 40.5), type = 'l')
}

