% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/PCA.R
\docType{methods}
\name{smoothScatter}
\alias{smoothScatter}
\alias{smoothScatter,PCA-method}
\title{generic for smoothScatter}
\usage{
smoothScatter(x, y = NULL, nbin = 128, bandwidth,
  colramp = colorRampPalette(c("white", blues9)), nrpoints = 100,
  ret.selection = FALSE, pch = ".", cex = 1, col = "black",
  transformation = function(x) x^0.25, postPlotHook = box, xlab = NULL,
  ylab = NULL, xlim, ylim, xaxs = par("xaxs"), yaxs = par("yaxs"), ...)

\S4method{smoothScatter}{PCA}(x, y = NULL, nbin = 128, bandwidth,
  colramp = colorRampPalette(c("white", blues9)), nrpoints = 100,
  ret.selection = FALSE, pch = ".", cex = 1, col = "black",
  transformation = function(x) x^0.25, postPlotHook = box, xlab = NULL,
  ylab = NULL, xlim, ylim, xaxs = par("xaxs"), yaxs = par("yaxs"), ...,
  comps = c(1, 2), pcType = "pcaScores", label = FALSE,
  labelThreshold = 1)
}
\arguments{
\item{x}{object of class PCA}

\item{y}{numeric usually NULL}

\item{nbin}{numeric}

\item{bandwidth}{numeric vector length 1 or 2}

\item{colramp}{numeric}

\item{nrpoints}{numeric}

\item{ret.selection}{logical}

\item{pch}{character}

\item{cex}{numeric}

\item{col}{character}

\item{transformation}{function}

\item{postPlotHook}{box}

\item{xlab}{NULL}

\item{ylab}{NULL}

\item{xlim}{numeric}

\item{ylim}{numeric}

\item{xaxs}{par}

\item{yaxs}{par}

\item{...}{additional args}

\item{comps}{numeric}

\item{pcType}{character}

\item{label}{boolean}

\item{labelThreshold}{numeric}
}
\value{
graphical output
}
\description{
generic for smoothScatter

smoothScatter method for PCA class
}
\examples{
library(tofsimsData)
data(tofsimsData)
testImage<-PCAnalysis(testImage, nComp = 4)
smoothScatter(analysis(testImage, 1), comps = c(1,2), 
pcType = 'pcaScores', xlab = 'comp 1', ylab = 'comp 2')
}

