% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNF.R
\name{SNR}
\alias{SNR}
\title{Signal-to-Noise Ratio (SNR)}
\usage{
SNR(stat, x, y)
}
\arguments{
\item{stat}{unknown}

\item{x}{unknown}

\item{y}{unknown}
}
\value{
matrix numeric with signal-to-noise ratios
}
\description{
SNR function for MNF to calculate Signal to Noise Ratio
}
\details{
function from mzimage to calculate signal-to-noise ratio function
}
