% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tofsims-package.R
\docType{package}
\name{tofsims-package}
\alias{tofsims-package}
\title{ToF-SIMS Toolbox (tofsims)}
\description{
ToF-SIMS Toolbox
}
\details{
\tabular{ll}{
Package: \tab tofsims\cr
Type: \tab Package\cr
Version: \tab 0.99.2\cr
Date: \tab 15-01-2016\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes \cr
}

Toolbox for Time-of-Flight Secondary Ion Mass-Spectrometry
(ToF-SIMS) data processing and analysis. The package facilitates 
importing of raw data files, loading preprocessed data and a range of 
multivariate analysis methods that are most commonly applied in the 
ToF-SIMS community.
}
\author{
Lorenz Gerber \email{lorenz.gerber@slu.se}

Viet Mai Hoang \email{hviet.0906@gmail.com}
}
\keyword{package}

