% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassSpectra.R
\docType{methods}
\name{unitMassPeaks}
\alias{unitMassPeaks}
\alias{unitMassPeaks,MassSpectra,numeric,numeric-method}
\title{Generic method for unitMassPeaks}
\usage{
unitMassPeaks(object, mzRange, widthAt, factor, upper = NULL, lower = NULL,
  ...)

\S4method{unitMassPeaks}{MassSpectra,numeric,numeric}(object, mzRange, widthAt,
  factor, upper = NULL, lower = NULL, ...)
}
\arguments{
\item{object}{object of class MassSpectra}

\item{mzRange}{vector numeric with lower and upper mass 
range limit for which to set unit mass peaks}

\item{widthAt}{vector numeric two mass values at which to sample for peak width}

\item{factor}{vector numeric two values summing up to 1 for 
setting assymetric peak width limits}

\item{upper}{vector numeric upper peak width limits}

\item{lower}{vectpr numeric lower peak width limits}

\item{...}{additional args}
}
\value{
object of class PeakList with unit mass peaks
}
\description{
Generic method for unitMassPeaks
}
\examples{
library(tofsimsData)
data(tofsimsData)
testSpectra <- calibPointNew(testSpectra, mz = 15, value = 15.01551)
testSpectra <- calibPointNew(testSpectra, mz = 181, value = 181.0228)
testSpectra <- recalibrate(testSpectra)
testSpectra <- unitMassPeaks(testSpectra, mzRange = c(1,200), widthAt = c(15, 181), 
factor = c(0.4, 0.6), lower = c(14.97, 15.05), upper = c(180.84, 181.43))
plot(testSpectra, mzRange = c(1,200), type = 'l')
}
