% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassDefinition.R
\name{validMassImageObject}
\alias{validMassImageObject}
\title{Validation method functionf for class MassImage objects}
\usage{
validMassImageObject(object)
}
\arguments{
\item{object}{object of class MassImage}
}
\value{
boolean class validity test
}
\description{
Validation method functionf for class MassImage objects
}
