% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassDefinition.R
\name{validMassSpectraObject}
\alias{validMassSpectraObject}
\title{Validation method function for class MassImage objects}
\usage{
validMassSpectraObject(object)
}
\arguments{
\item{object}{object of class MassSpectra}
}
\value{
boolean class validity test
}
\description{
Validation method function for class MassImage objects
}

