% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassDefinition.R
\name{validPCAObject}
\alias{validPCAObject}
\title{Validation method function for class PCA objects}
\usage{
validPCAObject(object)
}
\arguments{
\item{object}{object of class PCA}
}
\value{
boolean class validity test
}
\description{
Validation method function for class PCA objects
}

