% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClassDefinition.R
\name{validPeakListObject}
\alias{validPeakListObject}
\title{Validation method function for class PeakList objects}
\usage{
validPeakListObject(object)
}
\arguments{
\item{object}{object of class PeakList}
}
\value{
boolean class validity test
}
\description{
Validation method function for class PeakList objects
}

