% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGetters.R
\docType{methods}
\name{xdim,MassImage-method}
\alias{xdim,MassImage-method}
\title{Getter, method definition "xdim" on "MassImage"}
\usage{
\S4method{xdim}{MassImage}(object)
}
\arguments{
\item{object}{objet of class MassImage}
}
\value{
numeric x dimension of slot xy
}
\description{
Getter, method definition "xdim" on "MassImage"
}

