% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGetters.R
\docType{methods}
\name{xdim,PCA-method}
\alias{xdim,PCA-method}
\title{method xdim() for PCA class object}
\usage{
\S4method{xdim}{PCA}(object)
}
\arguments{
\item{object}{object of class PCA}
}
\value{
numeric x dimension of image
}
\description{
method xdim() for PCA class object
}
