% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{xdim<-}
\alias{xdim<-}
\title{generic setter method for "xdim"}
\usage{
xdim(object) <- value
}
\arguments{
\item{object}{object of class MassImage}

\item{value}{numeric x dimension of image}
}
\value{
object of class MassImage with updated x dimension
}
\description{
generic setter method for "xdim"
}

