% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{xdim}
\alias{xdim}
\title{generic accessor method for "xdim"}
\usage{
xdim(object)
}
\arguments{
\item{object}{object of class MassImage}
}
\value{
numeric value x dimension of mass image
}
\description{
generic accessor method for "xdim"
}
