% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/AllGetters.R,
%   R/AllSetters.R
\docType{methods}
\name{xy}
\alias{xy}
\alias{xy<-}
\alias{xy,MassImage-method}
\alias{xy<-,MassImage-method}
\title{\code{xy}, slot of \code{MassImage} class objects}
\usage{
xy(object)

xy(object) <- value

\S4method{xy}{MassImage}(object)

\S4method{xy}{MassImage}(object) <- value
}
\arguments{
\item{object}{object of class MassImage}

\item{value}{vector numeric two values for 
x and y dimension of image}
}
\value{
vector numeric with xy dimensions of image
}
\description{
\code{xy}, slot of \code{MassImage} class objects
}
\examples{
library(tofsimsData)
data(tofsimsData)
xy(testImage)
}
