% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R, R/MassImage.R
\docType{methods}
\name{xySpec}
\alias{xySpec}
\alias{xySpec,MassImage-method}
\title{Generic method \code{xySpec}}
\usage{
xySpec(object, x = NULL, y = NULL)

\S4method{xySpec}{MassImage}(object, x = NULL, y = NULL)
}
\arguments{
\item{object}{object of class MassImage}

\item{x}{numeric x coordinate from where to sample a mass spectra}

\item{y}{numeric y coordinate from where to sample a mass spectra}
}
\value{
object of class MassSpectra with selected mass spectra
}
\description{
Selection of Spectra

method xySpec extracts the mass spectra of positon x/y and puts
them in a MassSpectar class object
}
\details{
Selection of mass spectra by vectors of equal length for x and y.
}
\examples{
library(tofsimsData)
data(tofsimsData)
spectra100100<-xySpec(testImage, 100,100)
plot(spectra100100, type = 'l')
}
\author{
Lorenz Gerber <lorenz.gerber@slu.se>
}

