% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGetters.R
\docType{methods}
\name{ydim,MassImage-method}
\alias{ydim,MassImage-method}
\title{Getter, method definition "ydim" on "MassImage"}
\usage{
\S4method{ydim}{MassImage}(object)
}
\arguments{
\item{object}{object of class MassImage}
}
\value{
numeric y dimension of slot xy
}
\description{
Getter, method definition "ydim" on "MassImage"
}

