% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGetters.R
\docType{methods}
\name{ydim,PCA-method}
\alias{ydim,PCA-method}
\title{method ydim() for PCA class object}
\usage{
\S4method{ydim}{PCA}(object)
}
\arguments{
\item{object}{object of class PCA}
}
\value{
numeric y dimension of image
}
\description{
method ydim() for PCA class object
}

