% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{ydim<-}
\alias{ydim<-}
\title{generic setter method for "ydim"}
\usage{
ydim(object) <- value
}
\arguments{
\item{object}{object of class MassImage}

\item{value}{numeric y dimension of image}
}
\value{
updated object of type MassImage
}
\description{
generic setter method for "ydim"
}
