% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{ydim}
\alias{ydim}
\title{generic accessor method for "ydim"}
\usage{
ydim(object)
}
\arguments{
\item{object}{object of class MassImage}
}
\value{
numeric integer, y dimension of image
}
\description{
generic accessor method for "ydim"
}
