% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSpectra.R
\docType{methods}
\name{zdim,MassSpectra-method}
\alias{zdim,MassSpectra-method}
\title{method definition 'zdim' on 'MassSpectra'}
\usage{
\S4method{zdim}{MassSpectra}(object)
}
\arguments{
\item{object}{object of class MassSpectra}
}
\value{
numeric value
}
\description{
method definition 'zdim' on 'MassSpectra'
}

