% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenericDefinition.R
\name{zdim}
\alias{zdim}
\title{generic accessor method for "zdim"}
\usage{
zdim(object)
}
\arguments{
\item{object}{object of class MassImage}
}
\value{
numeric, number of mass channels / peaks
}
\description{
generic accessor method for "zdim"
}
