% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGetters.R
\docType{methods}
\name{resids,MCR-method}
\alias{resids,MCR-method}
\title{MCR accessor resids,}
\usage{
\S4method{resids}{MCR}(object)
}
\arguments{
\item{object}{object of class MCR}
}
\value{
resids from object
}
\description{
MCR accessor resids,
}

